+++ src/Plane/Point/PointInterface.php
@@ -2,14 +2,36 @@
 namespace samizdam\Geometry\Plane\Point;
 
 /**
- * 
- * @author samizdam
  *
+ * @author samizdam
+ *        
  */
 interface PointInterface
 {
 
+    /**
+     * Get angular (azimut at polar system)
+     *
+     * @return float
+     */
+    public function getAngular();
+
+    /**
+     * Get Polal Radius (distance to zero-point at polar system)
+     *
+     * @return float
+     */
+    public function getR();
+
+    /**
+     *
+     * @return float
+     */
     public function getX();
 
+    /**
+     *
+     * @return float
+     */
     public function getY();
 }