+++ CHANGELOG.md
@@ -18,15 +18,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Angle::getBisector().
 - PolygonInspector::isTangential(). 
 - PolygonInspector::isCiclic().   
-- Polar coordinates. 
+- Polar coordinates.  
 
 ### Changed
 - Normalize naming of arguments, fields and methods. To fix in some guideline.
 
 ## 0.5.4 - 2015-09-XX
 ### Added
-- create point by polar coordinates. 
-
+- Create point by polar coordinates. 
+- Getters for polar coordinates at Point. 
 
 ## 0.5.3 - 2015-09-26
 ### Added